requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiogram-media-group==0.5.1
aiohttp==3.8.6
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2023.7.22
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
psycopg2-binary==2.9.9
pydantic==2.3.0
pydantic-core==2.6.3
regex==2023.3.23
requests==2.28.2
tqdm==4.65.0
typing-extensions==4.7.1
urllib3==1.26.15
yarl==1.8.2
